+++ reflection/treeroot.go
@@ -356,7 +356,7 @@ func (ctx *ReflectionCtx) buildRootFromLargeUint(sourceType *ssztypes.TypeDescri
 		sourceValue = sourceValPtr.Elem()
 	}
 
-	sourceLen := uint32(sourceValue.Len())
+	sourceLen := int64(sourceValue.Len())
 	expectedLen := sourceType.Size / sourceType.ElemDesc.Size
 	if sourceLen > expectedLen {
 		return sszutils.ErrVectorLengthFn(sourceLen, expectedLen)
@@ -369,7 +369,7 @@ func (ctx *ReflectionCtx) buildRootFromLargeUint(sourceType *ssztypes.TypeDescri
 	isUint64 := sourceType.ElemDesc.Kind == reflect.Uint64
 	if isUint64 {
 		for i := 0; i < int(sourceType.Size/8); i++ {
-			if uint32(i) < sourceLen {
+			if int64(i) < sourceLen {
 				hh.AppendUint64(sourceValue.Index(i).Uint())
 			} else {
 				hh.AppendUint64(0)
@@ -635,7 +635,7 @@ func (ctx *ReflectionCtx) buildRootFromVector(sourceType *ssztypes.TypeDescripto
 	hashIndex := hh.StartTree(sszutils.TreeTypeBinary)
 
 	sliceLen := sourceValue.Len()
-	if uint32(sliceLen) > sourceType.Len {
+	if int64(sliceLen) > sourceType.Len {
 		if sourceType.Kind == reflect.Array {
 			sliceLen = int(vecLen)
 		} else {
@@ -644,7 +644,7 @@ func (ctx *ReflectionCtx) buildRootFromVector(sourceType *ssztypes.TypeDescripto
 	}
 
 	appendZero := 0
-	if uint32(sliceLen) < sourceType.Len {
+	if int64(sliceLen) < sourceType.Len {
 		appendZero = int(vecLen) - sliceLen
 	}
